refactor worker runners
